# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SAVVORA E-Commerce API",
    "description": "SAVVORA Production Enterprise E-Commerce Platform API",
    "version": "3.0.0"
  },
  "paths": {
...
    "/api/v1/admin/analytics": {
      "get": {
        "tags": [
          "8. Admin Dashboard & Analytics"
        ],
        "summary": "Get Sales Analytics",
        "description": "Fetches sales trends, revenue distribution, and growth metrics (Admin only).",
        "operationId": "get_sales_analytics_api_v1_admin_analytics_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/inventory/low-stock": {
      "get": {
        "tags": [
          "8. Admin Dashboard & Analytics"
        ],
        "summary": "Get Low Stock Inventory",
        "description": "Lists products with inventory below low stock threshold (Staff/Admin).",
        "operationId": "get_low_stock_inventory_api_v1_admin_inventory_low_stock_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/coupons": {
      "get": {
        "tags": [
          "8. Admin Dashboard & Analytics"
        ],
        "summary": "List Coupons",
        "description": "Lists all active and inactive discount coupons.",
        "operationId": "list_coupons_api_v1_admin_coupons_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CouponOut"
                  },
                  "title": "Response List Coupons Api V1 Admin Coupons Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "8. Admin Dashboard & Analytics"
        ],
        "summary": "Create Coupon",
        "description": "Creates a new discount coupon (Admin only).",
        "operationId": "create_coupon_api_v1_admin_coupons_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          },
          {
            "name": "discount_percent",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 10.0,
              "title": "Discount Percent"
            }
          },
          {
            "name": "max_discount",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 500.0,
              "title": "Max Discount"
            }
          },
          {
            "name": "min_order_value",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 999.0,
              "title": "Min Order Value"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CouponOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/tickets": {
      "get": {
        "tags": [
          "8. Admin Dashboard & Analytics"
        ],
        "summary": "List Support Tickets",
        "description": "Lists all customer support tickets.",
        "operationId": "list_support_tickets_api_v1_admin_tickets_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SupportTicketOut"
                  },
                  "type": "array",
                  "title": "Response List Support Tickets Api V1 Admin Tickets Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/api/coupons": {
      "get": {
        "tags": [
          "8. Coupon Management"
        ],
        "summary": "List Coupons",
        "operationId": "list_coupons_api_coupons_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "8. Coupon Management"
        ],
        "summary": "Create Coupon",
        "operationId": "create_coupon_api_coupons_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CouponCreatePayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CouponCreatePayload": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "discount_percent": {
            "type": "number",
            "title": "Discount Percent"
          },
          "max_discount": {
            "type": "number",
            "title": "Max Discount",
            "default": 500.0
          },
          "min_order_value": {
            "type": "number",
            "title": "Min Order Value",
            "default": 999.0
          },
          "usage_limit": {
            "type": "integer",
            "title": "Usage Limit",
            "default": 100
          }
        },
        "type": "object",
        "required": [
          "code",
          "discount_percent"
        ],
        "title": "CouponCreatePayload"
      },
      "CouponOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "discount_type": {
            "type": "string",
            "title": "Discount Type"
          },
          "discount_percent": {
            "type": "number",
            "title": "Discount Percent"
          },
          "max_discount": {
            "type": "number",
            "title": "Max Discount"
          },
          "min_order_value": {
            "type": "number",
            "title": "Min Order Value"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "discount_type",
          "discount_percent",
          "max_discount",
          "min_order_value",
          "is_active"
        ],
        "title": "CouponOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SupportTicketOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "ticket_number": {
            "type": "string",
            "title": "Ticket Number"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "priority": {
            "type": "string",
            "title": "Priority"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "ticket_number",
          "subject",
          "category",
          "priority",
          "status",
          "created_at"
        ],
        "title": "SupportTicketOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        }
      }
    }
  }
}